requirements.txt
click==6.7
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.15.0
opencv-python==3.4.2.17
Pillow==5.2.0
pytesseract==0.2.4
Werkzeug==0.14.1
